Add thumbnail version for image
Motivation:

To improve the user experience for everyone but especially
for users with low band width, there's now a smaller thumbnail
version displayed and a link to the full resolution image.

Modifications:

- Add thumbnail image with lower resolution.
- Show thumbnail version in blog post
- Add link to full resolution asset

Result:

The website will load quicker and use less band width while
visitors of the site are still able to inspect the full resolution
when using the link.