On very narrow screens, shorten long status names for 'filter by' string
With the addition of the upcoming feature flag filter button, the 'filter by' string can wrap to a second line on narrow mobile screens. The existing layout is not expecting multiple lines.

To address this the statuses 'Returned for Revision' and 'Scheduled for Review' are shortened to 'Returned' and 'Scheduled' in the 'filter by' string.

Note this only has an effect for screens less than 360 pixels wide and the status filter is of one and only one of those two statuses.