Removing Paul Hudson from Diversity in Swift
This removes my name from the list of Diversity in Swift workgroup members.

I had a great time working on the Diversity in Swift workgroup, and I am very grateful to the other members of the group for doing such a great job launching and supporting a range of initiatives. However, my two-year term has passed and I have now left the workgroup.

Given that the role of Swift workgroups is evolving (https://www.swift.org/blog/evolving-swift-project-workgroups/) this page may need further changes, but I'll leave that to folks who are more in the know.