Fix #394, aarch64 download links broken
Fixed aarch64 download links for the current stable release.

Motivation:

Currently the `x86_64` and `aarch64` downloads for the current release point both to the `x86_64` download artifact for the release.

Modifications:

Added a condition in the download template which appends the arch to the url if it's not the default `x86_64`

Result:

Download links for `aarch64` will point to `aarch64` downlaod artifacts and `x86_64` to the `x86_64` ones.

Signed-off-by: paulober <44974737+paulober@users.noreply.github.com>