Fixed #swift-evolution menu item and 4 more links
This commit fixes 4 more broken links: the navigation menu item,
the reference-less link, and two other blog posts.

The `_navigation.yml` and several pages linked to
`/contributing/#evolution-process`, and we had an `<a>` tag with
`name="evolution-process"`, which is invalid in `htmlproofer`s eyes
because it's a link that does not have a reference.

Jekyll makes `id`s for every heading automatically, and we have
`/contributing/#swift-evolution` that we should link to instead.