Add an invitation to contribute to TSPL below its docs links.
Motivation:

Someone suggested that such an invitation in this part of the
site might make it more obvious that we're open to contributions.

Modifications:

Added a link and encouragement to contribute below the links
to the Swift book on documentation/index.md.

Result:

Readers are more aware that the book is open to contributions.