Link directly to package repositories
This change updates the new Packages page to link directly to package repositories rather than the Swift Package Index. A new section has been added to display secondary links, where a link to the Swift Package Index now resides. swiftinit.org is one such alternative index that could be added in a future PR.

Linking directly to the Swift Package Index is problematic for a few reasons:
1. It forms a 1-1 relationship between swift.org and the Swift Package Index, which excludes the option for other community sites to be represented. This tight relationship also discourages the development of alternative indexes, as the Swift Package Index is implied to be the "official" index.
2. People accustomed to visiting repositories directly may be unfamiliar with the Swift Package Index, which leads to a confusing experience.
3. The Swift Package Index deprives package authors of GitHub stars and potentially even sponsors because it does not display the appropriate links. Instead, it largely duplicates the content already present on the source repository and thus reduces the likelihood that a visitor clicks through to the repository. Perhaps controversially, the Swift Package Index instead displays its own prominent Supporters link.

This change does not change the behavior of the "Browse more packages" button, which currently links to the Swift Package Index. Perhaps a follow-up PR could similarly change this feature to display multiple indexes.