Stop using non-standard search event evolution dashboard
The next version of Safari removes the non-standard search event.
This causes the search field on the evolution dashboard to stop working.
The issue can be reproduced in Safari Technology Preview and Firefox.

This fix makes two changes:
- Uses the standard 'input' event instead of 'search' event
- Removes the non-standard 'incremental' attribute

The input event triggers the event listener on each change of input.

The behavior with this fix is unchanged to the behavior in the current Safari version.

Tested on macOS 14.1 in Safari, Chrome, Firefox, and
Safari Technology Preview.

Fixes #440