Use 'current version' instead of specific Xcode version
The Xcode version is read from the last entry of swift_releases.yml.
Multiple Xcode versions can ship with the same Swift version, so
the Xcode version displayed can be out of date.

The exact Xcode version is not very important to display on the page.
So, instead of creating and maintaining an Xcode version data file,
the text changes to say 'current version'.

Since the user will clicking through to the Mac App Store, they will
see the correct version number on the app page in the store.

This fixes #493.