Contributor Experience Workgroup: Add links to members' GitHub accounts (#621)
Currently, the swift repository docs
[are asking](https://github.com/apple/swift/blob/ee24bdf06f59d05fea996c230eb4024aa04dd4f5/docs/HowToGuides/FirstPullRequest.md#:~:text=message%20the%20Contributor%20Experience%20Workgroup)
contributors to message the workgroup on the forums for help, which is
a drag when you can simply reach out to a member in that same issue or
pull request. Filling out GitHub accounts will enable us to direct
people to the charter's member list instead.

Also, reformat the member list into a table, which is easier to visually
parse with links to multiple accounts.