[Articles] Getting started setting up Neovim for Swift  (#675)
* Zero to Swift: Getting Started with Neovim

This is an article on setting up Neovim for Swift development.
It covers setting up a package manager, LSP, auto-completions, snippets,
and finally some helpful autocommands.

* Editing and rewording

Responding to editorial feedback, rewording some things, and removing
the autocmds bit. The autocmds-bits were mostly talked about things that
matter to compiler engineers, but aren't really necessary for someone
using the language normally.

* Added lead-in sentence for examples.

* Spelling lazy.nvim

Changing textual instances of "Lazy" to _lazy.nvim_ to match the name of
the plugin in the plugin documentation.

* Add jekyll header to nvim document

* Change shell code-blocks to Bourne Again SHell

* Move plugin links into document

* Nit to qualify the vi editor

* Improve the introduction

* Include link to Swift install page

* Edited intro only for present tense, acronyms, and scannability.

* Called out intro caveats as "Tip" and "Note"

* Use console for shell codeblocks

The console code-block language recognizes the prompt. With recent
changes to the CSS, things marked as part of the prompt aren't
selectable making it easier to copy-paste commands.

* Italicize LuaSnip and nvm-cmp

These are the names of the plugins, replacing backticks with italicizing
the plugin name.

* Use title of GitHub issues instead of raw links

It's a little easier to understand what is going on looking at the title
of the issues about dynamic registration and neovim than the raw link
text.

* Replace Swift install with swift.org reference

Removing the section on downloading the tarball and replacing it with a
reference back to the swift.org installation page.

* Capitalize Vim and Vi

Fixing capitalization on Vim and Vi. Also removing the expansion of Vim.
It's the name of the tool at this point and folks reading this document
should be familiar with it, or at the very least, "VIsual IMproved", is
not particularly enlightening if you don't know what it is.

* Reword Neovim install

Moving the link to the Neovim install page up for that folks who are not
using Ubuntu.

* Splitting up longer paragraphs

Answering some review feedback to split up some of the longer
paragraphs.

---------

Co-authored-by: RobinBateman808 <157847969+RobinBateman808@users.noreply.github.com>