Modernizes visual appearance of tabs
Updates the .interactive-tabs class to take a more layered appearance.
Hover and selection states are visually distinct and rely on the
existing color palette for layers.

Adds support for "prefers-contrast: more" to use the previous higher
contrast style.

Moves the css for tabs to a common location so it can be easily found
and reused between pages.