A new Member contributor role that can trigger CI
We currently have three roles that contributors to the Swift project on GitHub can have: Everybody can participate the Swift open source project and become a *Contributor*. After creating a track record by creating 5 non-trivial pull requests that were accepted without modifications, contributors can request [*Commit Access*](https://www.swift.org/contributing/#commit-access), which allows them to merge PRs and also trigger CI. Finally, contributors can become [*Code Owners*](https://www.swift.org/community/#code-owners).

With the current structure, gaining the ability to trigger CI has quite a high bar and hinders new contributors from accelerating their contributions to Swift. We are thus introducing a new role: After contributing to Swift multiple times, contributors can request to become a *Member*, which grants them the ability to trigger CI but does not give them commit access yet. This way they can test changes on their own schedule and taking the burden to trigger CI off those with commit access.

For now, this is not indented to change any of the other roles. We are working on improvements and clarifications to the other roles as well but would like to keep this change focused on the *Member* role.

Finally a word on the role’s naming: Becoming a *Member* will be implemented by adding the user to the `swiftlang` organization on GitHub, which also displays a *Member* badge next in PRs opened by that user. We want to align the naming in Swift’s contributor ladder with that of GitHub to avoid confusion.