Remove line breaks from long WinGet command
The WinGet command for installing Windows platform dependencies
contained line breaks. When copied, the line breaks would break the
command up in to multiple lines. The first line is a valid command and
so runs sucessfully but the subsequent lines are truncated and will
produce an error if run.

Update the command to remove line breaks, and update the css for
`.code-box pre code`` to add back in line breaks in css that are not
preserved when the command is copied.

The swiftly install command breaks the command up with backslashes `\`
but on Windows we would have to either use `^` for cmd.exe or backtick
for powershell. This ambiguity means its safer just to put the whole
command on one line.